feat: Add user approval/rejection endpoints with fixes
- Added approve transaction endpoint (POST /api/transaction/approve)
- Added reject transaction endpoint (POST /api/transaction/reject)
- Added pending transactions list endpoint (GET /api/transactions/pending)
- Updated CSV structure with User_Action, Actioned_By, Action_Timestamp, Action_Comments
- Added update_transaction_status helper function
- Fixed Decimal to float conversion in db_service.py
- Fixed NaN/Infinity handling in pending transactions endpoint
- Updated field names: approved_by/rejected_by -> customer_id (customer self-service)
- Updated Postman collection with new endpoints

+++ backend/db_service.py
@@ -182,9 +182,9 @@ class DatabaseService:
                     "current_month_spending": 0.0
                 }
             
-            avg_amount = df['AmountInAed'].mean()
-            std_amount = df['AmountInAed'].std() if len(df) > 1 else 2000.0
-            max_amount = df['AmountInAed'].max()
+            avg_amount = float(df['AmountInAed'].mean())
+            std_amount = float(df['AmountInAed'].std()) if len(df) > 1 else 2000.0
+            max_amount = float(df['AmountInAed'].max())
             txn_count = len(df)
             
             intl_ratio = 0.0
